Mouse device fully implemented
Mouse device has been fully implemented in terms of the rust class. Events are tracking state of device. Pending testing. 

Adding the empty_device file. Not much is in this file currently but the purpose of this is to be able to parse json information from the device and create the tuple.